requirements.txt
asgiref==3.4.1
Babel==2.9.1
crispy-bootstrap5==0.4
Django==3.2.5
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-phonenumber-field==5.2.0
phonenumbers==8.12.28
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
